@@ -1,50 +1,50 @@
 source 'https://rubygems.org'
 
+git_source(:github) do |repo_name|
+  repo_name = "#{repo_name}/#{repo_name}" unless repo_name.include?("/")
+  "https://github.com/#{repo_name}.git"
+end
+
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '5.0.2'
-
-gem 'pg', '~> 0.19.0'
-
+gem 'rails', '~> 5.1.2'
+# Use sqlite3 as the database for Active Record
+gem 'sqlite3'
+# Use Puma as the app server
+gem 'puma', '~> 3.7'
 # Use SCSS for stylesheets
 gem 'sass-rails', '~> 5.0'
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
-# Use CoffeeScript for .coffee assets and views
-gem 'coffee-rails', '~> 4.1.0'
 # See https://github.com/rails/execjs#readme for more supported runtimes
 # gem 'therubyracer', platforms: :ruby
 
-# Use jquery as the JavaScript library
-gem 'jquery-rails'
-# Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
-gem 'turbolinks'
+# Use CoffeeScript for .coffee assets and views
+gem 'coffee-rails', '~> 4.2'
+# Turbolinks makes navigating your web application faster. Read more: https://github.com/turbolinks/turbolinks
+gem 'turbolinks', '~> 5'
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.0'
-# bundle exec rake doc:rails generates the API under doc/api.
-gem 'sdoc', '~> 0.4.0', group: :doc
-
+gem 'jbuilder', '~> 2.5'
+# Use Redis adapter to run Action Cable in production
+# gem 'redis', '~> 3.0'
 # Use ActiveModel has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
-
-# Use Unicorn as the app server
-# gem 'unicorn'
 
 # Use Capistrano for deployment
 # gem 'capistrano-rails', group: :development
 
-gem 'sufia', '7.3.1'
-
-gem 'blacklight_advanced_search', '~> 6.0'
-
-
 group :development do
-  # Access an IRB console on exception pages or by using <%= console %> in views
-  gem 'web-console', '~> 2.0'
+  # Access an IRB console on exception pages or by using <%= console %> anywhere in the code.
+  gem 'web-console', '>= 3.3.0'
+  gem 'listen', '>= 3.0.5', '< 3.2'
   # helps enforce good coding standards
   gem 'rubocop', '~> 0.49.1', require: false
 end
 
+# Windows does not include zoneinfo files, so bundle the tzinfo-data gem
+gem 'tzinfo-data', platforms: [:mingw, :mswin, :x64_mingw, :jruby]
+gem 'hyrax', github: 'samvera/hyrax'#, :tag => 'v1.0.3'
+gem 'blacklight_advanced_search', '~> 6.0'
 
 group :development, :test do
   gem 'solr_wrapper', '>= 0.3'
@@ -53,21 +53,20 @@ group :development, :test do
   gem 'webmock', '~> 3.0.1'
   gem 'vcr', '~> 3.0.3'
   gem 'factory_girl_rails', '~> 4.8.0'
-  gem 'byebug'
+  # Call 'byebug' anywhere in the code to stop execution and get a debugger console
+  gem 'byebug', platforms: [:mri, :mingw, :x64_mingw]
   gem 'json-schema', '~> 2.8.0'
   gem 'rails-controller-testing', '~> 1.0.2'
+  # Adds support for Capybara system testing and selenium driver
+  gem 'capybara', '~> 2.13'
+  gem 'selenium-webdriver'
 end
 
-gem 'rsolr', '~> 1.0'
+gem 'rsolr', '>= 1.0'
+gem 'jquery-rails'
 gem 'devise'
-gem 'devise-guests', '~> 0.5'
+gem 'devise-guests', '~> 0.6'
 gem 'hydra-role-management'
-gem 'aws-sdk', '~> 2.9.15'
-
-gem 'puma', '~> 3.8.2'
-gem 'tinymce-rails-imageupload', '4.0.17.beta.2' # we need this specific version to fix an asset compilation problem
-
-# override this for a local folder with the command:
-# $ bundle config local.willow_sword $APP_HOME/willow_sword
-gem 'willow_sword', git: "https://github.com/cottagelabs/willow_sword.git", branch: "master"
 gem 'bootstrap-datepicker-rails'
+gem 'aws-sdk', '>= 2.9.15'
+gem 'tinymce-rails-imageupload', '4.0.17.beta.2' # we need this specific version to fix an asset compilation problem
