@@ -1,3 +1,3 @@
-class Qa::LocalAuthorityEntry < ActiveRecord::Base
+class Qa::LocalAuthorityEntry < ApplicationRecord
   belongs_to :local_authority
 end
