@@ -1,16 +1,13 @@
-class User < ActiveRecord::Base
+class User < ApplicationRecord
   # Connects this user object to Hydra behaviors.
   include Hydra::User
   # Connects this user object to Role-management behaviors.
   include Hydra::RoleManagement::UserRoles
 
 
-  # Connects this user object to Curation Concerns behaviors.
-  include CurationConcerns::User
-  # Connects this user object to Sufia behaviors.
-  include Sufia::User
-  include Sufia::UserUsageStats
-
+  # Connects this user object to Hyrax behaviors.
+  include Hyrax::User
+  include Hyrax::UserUsageStats
 
 
 
