@@ -4,16 +4,11 @@ class ApplicationController < ActionController::Base
   include Blacklight::Controller
   include Hydra::Controller::ControllerBehavior
 
-  # Adds CurationConcerns behaviors to the application controller.
-  include CurationConcerns::ApplicationControllerBehavior  
-  # Adds Sufia behaviors into the application controller 
-  include Sufia::Controller
-
-  include CurationConcerns::ThemedLayoutController
+  # Adds Hyrax behaviors into the application controller
+  include Hyrax::Controller
+  include Hyrax::ThemedLayoutController
   with_themed_layout '1_column'
 
 
-  # Prevent CSRF attacks by raising an exception.
-  # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
 end
