@@ -1,9 +1,7 @@
 class Ability
   include Hydra::Ability
   
-  include CurationConcerns::Ability
-  include Sufia::Ability
-
+  include Hyrax::Ability
   self.ability_logic += [:everyone_can_create_curation_concerns]
 
   # Define any customized permissions here.
