@@ -1,8 +1,7 @@
-# Generated by curation_concerns:models:install
+# Generated by hyrax:models
 class Collection < ActiveFedora::Base
-  include ::CurationConcerns::CollectionBehavior
+  include ::Hyrax::CollectionBehavior
   # You can replace these metadata if they're not suitable
-  include CurationConcerns::BasicMetadata
-
-  scope :ordered, -> {order("id ASC")}
+  include Hyrax::BasicMetadata
+  self.indexer = Hyrax::CollectionWithBasicMetadataIndexer
 end
