@@ -11,13 +11,13 @@ class PersonStatement < ActiveTriples::Resource
   property :affiliation, predicate: ::RDF::Vocab::VMD.affiliation
 
   ## Necessary to get AT to create hash URIs.
-  def initialize(uri, parent)
-    if uri.try(:node?)
-      uri = RDF::URI("#person#{uri.to_s.gsub('_:', '')}")
-    elsif uri.start_with?("#")
-      uri = RDF::URI(uri)
-    end
-    super
-  end
+  #def initialize(uri, parent)
+  #  if uri.try(:node?)
+  #    uri = RDF::URI("#person#{uri.to_s.gsub('_:', '')}")
+  #  elsif uri.start_with?("#")
+  #    uri = RDF::URI(uri)
+  #  end
+  #  super
+  #end
 
 end
