@@ -10,13 +10,13 @@ class RelationStatement < ActiveTriples::Resource
   property :relationship_role, predicate: ::RDF::Vocab::MODS.roleRelationshipRole
 
   ## Necessary to get AT to create hash URIs.
-  def initialize(uri, parent)
-    if uri.try(:node?)
-      uri = RDF::URI("#relation#{uri.to_s.gsub('_:', '')}")
-    elsif uri.start_with?("#")
-      uri = RDF::URI(uri)
-    end
-    super
-  end
+  # def initialize(uri, parent)
+  #   if uri.try(:node?)
+  #     uri = RDF::URI("#relation#{uri.to_s.gsub('_:', '')}")
+  #   elsif uri.start_with?("#")
+  #     uri = RDF::URI(uri)
+  #   end
+  #   super
+  # end
 
 end
