@@ -8,13 +8,13 @@ class RightsStatement < ActiveTriples::Resource
   property :start_date, predicate: ::RDF::Vocab::DISCO.startDate
 
   ## Necessary to get AT to create hash URIs.
-  def initialize(uri, parent)
-    if uri.try(:node?)
-      uri = RDF::URI("#rights#{uri.to_s.gsub('_:', '')}")
-    elsif uri.start_with?("#")
-      uri = RDF::URI(uri)
-    end
-    super
-  end
+  # def initialize(uri, parent)
+  #   if uri.try(:node?)
+  #     uri = RDF::URI("#rights#{uri.to_s.gsub('_:', '')}")
+  #   elsif uri.start_with?("#")
+  #     uri = RDF::URI(uri)
+  #   end
+  #   super
+  # end
 
 end
