@@ -1,5 +1,4 @@
-# Generated by curation_concerns:models:install
+# Generated by hyrax:models:install
 class FileSet < ActiveFedora::Base
-  include ::CurationConcerns::FileSetBehavior
-  include Sufia::FileSetBehavior
+  include ::Hyrax::FileSetBehavior
 end
