@@ -1,8 +1,9 @@
 # frozen_string_literal: true
 class SearchBuilder < Blacklight::SearchBuilder
   include Blacklight::Solr::SearchBuilderBehavior
+  # Add a filter query to restrict the search to documents the current user has access to
   include Hydra::AccessControlsEnforcement
-  include CurationConcerns::SearchFilters
+  include Hyrax::SearchFilters
 
 
   ##
