@@ -1,5 +1,4 @@
-# app/renderers/email_attribute_renderer.rb
-class NestedDateAttributeRenderer < CurationConcerns::Renderers::DateAttributeRenderer
+class NestedDateAttributeRenderer < Hyrax::Renderers::DateAttributeRenderer
   private
   def attribute_value_to_html(value)
     value = JSON.parse(value)
