@@ -1,7 +1,4 @@
 Rails.application.routes.draw do
-  Hydra::BatchEdit.add_routes(self)
-  mount Qa::Engine => '/authorities'
-
   
   mount Blacklight::Engine => '/'
   
@@ -14,15 +11,11 @@ Rails.application.routes.draw do
   devise_for :users
   mount Hydra::RoleManagement::Engine => '/'
 
-  # Sufia 7.3 - mount sufia before cc
-  mount Sufia::Engine => '/'
-
-  mount CurationConcerns::Engine, at: '/'
+  mount Qa::Engine => '/authorities'
+  mount Hyrax::Engine, at: '/'
   resources :welcome, only: 'index'
-  root 'sufia/homepage#index'
-  curation_concerns_collections
+  root 'hyrax/homepage#index'
   curation_concerns_basic_routes
-  curation_concerns_embargo_management
   concern :exportable, Blacklight::Routes::Exportable.new
 
   resources :solr_documents, only: [:show], path: '/catalog', controller: 'catalog' do
@@ -37,62 +30,5 @@ Rails.application.routes.draw do
     end
   end
 
-  mount WillowSword::Engine, at: "/sword"
-  
-  # The priority is based upon order of creation: first created -> highest priority.
-  # See how all your routes lay out with "rake routes".
-
-  # You can have the root of your site routed with "root"
-  # root 'sufia/homepage#index'
-
-  # Example of regular route:
-  #   get 'products/:id' => 'catalog#view'
-
-  # Example of named route that can be invoked with purchase_url(id: product.id)
-  #   get 'products/:id/purchase' => 'catalog#purchase', as: :purchase
-
-  # Example resource route (maps HTTP verbs to controller actions automatically):
-  #   resources :products
-
-  # Example resource route with options:
-  #   resources :products do
-  #     member do
-  #       get 'short'
-  #       post 'toggle'
-  #     end
-  #
-  #     collection do
-  #       get 'sold'
-  #     end
-  #   end
-
-  # Example resource route with sub-resources:
-  #   resources :products do
-  #     resources :comments, :sales
-  #     resource :seller
-  #   end
-
-  # Example resource route with more complex sub-resources:
-  #   resources :products do
-  #     resources :comments
-  #     resources :sales do
-  #       get 'recent', on: :collection
-  #     end
-  #   end
-
-  # Example resource route with concerns:
-  #   concern :toggleable do
-  #     post 'toggle'
-  #   end
-  #   resources :posts, concerns: :toggleable
-  #   resources :photos, concerns: :toggleable
-
-  # Example resource route within a namespace:
-  #   namespace :admin do
-  #     # Directs /admin/products/* to Admin::ProductsController
-  #     # (app/controllers/admin/products_controller.rb)
-  #     resources :products
-  #   end
-
-  Hydra::BatchEdit.add_routes(self)
+  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end
