@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-describe Collection do
+RSpec.describe Collection do
   it "has tests" do
     skip "Add your tests here"
   end
